requirements.txt
accelerate==0.25.0
addict==2.4.0
aiohappyeyeballs==2.6.1
aiohttp==3.13.3
aiosignal==1.4.0
alembic==1.18.4
aliyun-python-sdk-core==2.16.0
aliyun-python-sdk-kms==2.16.5
annotated-doc==0.0.4
annotated-types==0.7.0
anyio==4.12.1
appnope==0.1.4
asteroid-filterbanks==0.4.0
asttokens==3.0.1
async-timeout==5.0.1
attrs==25.4.0
audioread==3.1.0
av==16.1.0
backports.functools-lru-cache==2.0.0
certifi==2026.1.4
cffi==2.0.0
charset-normalizer==3.4.4
click==8.3.1
coloredlogs==15.0.1
colorlog==6.10.1
comm==0.2.3
contourpy==1.3.2
crcmod==1.7
cryptography==46.0.3
ctranslate2==4.6.3
cycler==0.12.1
datasets==2.10.1
debugpy==1.8.19
decorator==5.2.1
dill==0.3.5.1
distro==1.9.0
einops==0.8.1
exceptiongroup==1.3.1
executing==2.2.1
fastapi==0.128.0
faster-whisper==1.2.1
filelock==3.20.3
flatbuffers==25.12.19
fonttools==4.61.1
frozenlist==1.8.0
fsspec==2023.10.0
gast==0.7.0
google-ai-generativelanguage==0.6.15
google-api-core==2.30.0
google-api-python-client==2.190.0
google-auth==2.49.0.dev0
google-auth-httplib2==0.3.0
google-genai==1.65.0
googleapis-common-protos==1.72.0
grpcio==1.78.0
grpcio-status==1.71.2
h11==0.16.0
hf-xet==1.2.0
httpcore==1.0.9
httplib2==0.31.2
httpx==0.28.1
huggingface-hub==0.23.5
humanfriendly==10.0
HyperPyYAML==1.2.3
idna==3.11
importlib_metadata==8.7.1
indic-transliteration==2.3.75
ipykernel==7.1.0
ipython==8.38.0
jedi==0.19.2
Jinja2==3.1.6
jmespath==0.10.0
joblib==1.5.3
julius==0.2.7
jupyter_client==8.8.0
jupyter_core==5.9.1
kiwisolver==1.4.9
lazy_loader==0.4
librosa==0.11.0
lightning==2.6.1
lightning-utilities==0.15.2
llvmlite==0.46.0
Mako==1.3.10
markdown-it-py==4.0.0
MarkupSafe==3.0.3
matplotlib==3.10.8
matplotlib-inline==0.2.1
mdurl==0.1.2
mlx==0.31.0
mlx-metal==0.31.0
mlx-whisper==0.4.3
modelscope==1.9.5
more-itertools==10.8.0
mpmath==1.3.0
msgpack==1.1.2
multidict==6.7.0
multiprocess==0.70.13
nest-asyncio==1.6.0
networkx==3.4.2
nltk==3.9.2
numba==0.63.1
numpy==1.24.4
ollama==0.6.1
onnxruntime==1.23.2
opentelemetry-api==1.39.1
opentelemetry-exporter-otlp==1.39.1
opentelemetry-exporter-otlp-proto-common==1.39.1
opentelemetry-exporter-otlp-proto-grpc==1.39.1
opentelemetry-exporter-otlp-proto-http==1.39.1
opentelemetry-proto==1.39.1
opentelemetry-sdk==1.39.1
opentelemetry-semantic-conventions==0.60b1
optuna==4.7.0
oss2==2.19.1
packaging==26.0
pandas==2.3.3
parso==0.8.5
pexpect==4.9.0
pillow==12.1.0
platformdirs==4.5.1
pooch==1.8.2
primePy==1.3
prompt_toolkit==3.0.52
propcache==0.4.1
proto-plus==1.27.1
protobuf==5.29.6
psutil==7.2.1
ptyprocess==0.7.0
pure_eval==0.2.3
pyannote-database==6.1.1
pyannote-pipeline==4.0.0
pyannoteai-sdk==0.4.0
pyarrow==23.0.0
pyarrow-hotfix==0.7
pyasn1==0.6.2
pyasn1_modules==0.4.2
PyAudio==0.2.14
pycparser==3.0
pycryptodome==3.23.0
pydantic==2.12.5
pydantic_core==2.41.5
Pygments==2.19.2
pyparsing==3.3.2
python-dateutil==2.9.0.post0
python-dotenv==1.2.1
pytorch-lightning==2.6.1
pytorch-metric-learning==2.9.0
pytz==2025.2
PyYAML==6.0.3
pyzmq==27.1.0
regex==2026.1.15
requests==2.32.5
Resemblyzer==0.1.4
responses==0.18.0
rich==14.3.1
roman==5.2
ruamel.yaml==0.18.17
ruamel.yaml.clib==0.2.15
safetensors==0.7.0
scikit-learn==1.7.2
scipy==1.15.3
sentence-transformers==2.6.1
sentencepiece==0.2.1
shellingham==1.5.4
sherpa-onnx-core==1.12.23
sherpa_onnx==1.12.23
silero-vad==6.2.0
simplejson==3.20.2
six==1.17.0
sniffio==1.3.1
sortedcontainers==2.4.0
sounddevice==0.5.5
soundfile==0.13.1
soxr==1.0.0
speechbrain==0.5.16
SQLAlchemy==2.0.46
srt==3.5.3
stack-data==0.6.3
starlette==0.50.0
sympy==1.14.0
tenacity==9.1.4
threadpoolctl==3.6.0
tiktoken==0.12.0
tokenizers==0.15.2
toml==0.10.2
tomli==2.4.0
torch==2.1.2
torch-audiomentations==0.12.0
torch_pitch_shift==1.2.5
torchaudio==2.1.2
torchcodec==0.10.0
torchmetrics==1.8.2
tornado==6.5.4
tqdm==4.67.1
traitlets==5.14.3
transformers==4.35.2
typer==0.21.1
typer-slim==0.21.1
typing==3.7.4.3
typing-inspection==0.4.2
typing_extensions==4.15.0
tzdata==2025.3
uritemplate==4.2.0
urllib3==2.6.3
uvicorn==0.40.0
vosk==0.3.44
wcwidth==0.3.1
webrtcvad==2.0.10
websockets==16.0
xxhash==3.6.0
yapf==0.43.0
yarl==1.22.0
zipp==3.23.0
